Create decentralized lottery
Starting a lottery (startLottery):
The manager (contract creator) can start a lottery by setting the lotteryActive flag to true. This is a signal to participants that funds can be accepted.

Participating in a lottery (enter):
Any user can enter the lottery by sending at least 0.01 ETH. The participant's address is stored in the players array.

Ending a lottery and choosing a winner (endLottery):
After funds are accepted, the manager calls the endLottery function. The contract generates a pseudo-random number using the current block parameters, calculates the winner's index, and transfers the entire contract balance to the winner's address. The list of participants is then cleared in preparation for the next round.

Getting the number of participants (getPlayersCount):
Allows you to find out the current number of participants in an active lottery.